// frontend-recode/src/types/impl/api/static/operator.ts
export type Operator = {
    id?: string;
    name: string;
// ...
    phases: {
        characterPrefabKey: string;
        rangeId: string;
        maxLevel: number;
        attributesKeyFrames: {
            level: number;
            data: {
                maxHp: number;
// ...
            };
        }[];
        evolveCost:
            | {
                  id: string;
                  count: number;
                  type: ItemType;
              }[]
            | null;
    }[];
// ...
};

export enum OperatorPosition {
    RANGED = "RANGED",
    MELEE = "MELEE",
}

export enum OperatorPhase {
    ELITE_0 = "PHASE_0",
    ELITE_1 = "PHASE_1",
    ELITE_2 = "PHASE_2",
}

export enum OperatorRarity {
    sixStar = "TIER_6",
    fiveStar = "TIER_5",
    fourStar = "TIER_4",
    threeStar = "TIER_3",
    twoStar = "TIER_2",
    oneStar = "TIER_1",
}

export enum OperatorProfession {
    MEDIC = "MEDIC",
    CASTER = "CASTER",
    GUARD = "WARRIOR",
    VANGUARD = "PIONEER",
    SNIPER = "SNIPER",
    SPECIALIST = "SPECIAL",
    SUPPORTER = "SUPPORT",
    DEFENDER = "TANK",
}